#61 - 코드 리팩토링 및 에러 메시지 변경

+++ src/main/java/com/fastcampus/projectboard/Service/UserService.java
@@ -9,8 +9,6 @@ import com.fastcampus.projectboard.repository.SaveFileRepository;
 import com.fastcampus.projectboard.repository.UserAccountRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
@@ -21,7 +19,6 @@ import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.multipart.MultipartFile;
 
 import javax.persistence.EntityNotFoundException;
 import java.io.IOException;
@@ -46,48 +43,32 @@ public class UserService {
 
 
 
-    public String saveUserAccount(UserAccount.SignupDto user, SaveFile.FileDto fileDto) throws IOException {
-        String password = user.getPassword1();
+    public String saveUserAccount(UserAccount.SignupDto user, SaveFile.SaveFileDto saveFileDto) throws IOException {
         UserAccount account = userAccountRepository.save(user.toEntity());
-        account.setUserPassword(new BCryptPasswordEncoder().encode(password));
-        account.setProfileImg(fileDto.toEntity());
+        account.setUserPassword(new BCryptPasswordEncoder().encode(user.getPassword1()));
+        account.setProfileImg(saveFileDto.toEntity());
         return user.getUserId();
     }
 
-    public SaveFile.FileDto changeAccountProfileImg(String id, SaveFile.FileDto fileDto) throws IOException {
-        UserAccount account = userAccountRepository.findById(id).orElseThrow(()->new IllegalArgumentException("해당 유저가 없습니다."));
+    public SaveFile.SaveFileDto changeAccountProfileImg(String id, SaveFile.SaveFileDto saveFileDto) throws IOException {
+        UserAccount account = userAccountRepository.findById(id).orElseThrow(EntityNotFoundException::new);
         if(account.getProfileImg()!=null && !account.getProfileImg().getFileSize().equals(0L)){
-            FileUtil.deleteFile(SaveFile.FileDto.from(account.getProfileImg()));
+            FileUtil.deleteFile(SaveFile.SaveFileDto.from(account.getProfileImg()));
         }
-        SaveFile.FileDto currentImg = SaveFile.FileDto.from(account.getProfileImg());
-        account.setProfileImg(fileDto.toEntity());
+        SaveFile.SaveFileDto currentImg = SaveFile.SaveFileDto.from(account.getProfileImg());
+        account.setProfileImg(saveFileDto.toEntity());
         return currentImg;
     }
-
-
-
 
     @Transactional(readOnly = true)
     public boolean isUserExists(String userId){
         return userAccountRepository.existsById(userId);
     }
 
-
-
-    // 토큰 재발급 관련 메서드
-
-
-    // 권한 가져오기
-    @Transactional(readOnly = true)
-    public String getAuthorities(Authentication authentication) {
-        return authentication.getAuthorities().stream()
-                .map(GrantedAuthority::getAuthority)
-                .collect(Collectors.joining(","));
-    }
     public void updateUserAccount(String userId,UserAccount.UserAccountUpdateRequestDto userDto) {
         try {
             PasswordEncoder passwordEncoder = new BCryptPasswordEncoder();
-            UserAccount userAccount = userAccountRepository.findById(userId).orElseThrow(()->new EntityNotFoundException("해당 유저가 없습니다."));
+            UserAccount userAccount = userAccountRepository.findById(userId).orElseThrow(EntityNotFoundException::new);
             if (userDto.password1() != null) { userAccount.setUserPassword(passwordEncoder.encode(userDto.password1())); }
             if (userDto.nickname() != null) { userAccount.setNickname(userDto.nickname()); }
             if (userDto.email() != null) { userAccount.setEmail(userDto.email()); }
@@ -100,7 +81,7 @@ public class UserService {
     @Transactional(readOnly = true)
     public UserAccount.UserAccountDto getUserAccount(String userId) {
         try {
-            UserAccount userAccount = userAccountRepository.findById(userId).orElseThrow(()->new EntityNotFoundException("해당 유저가 없습니다."));
+            UserAccount userAccount = userAccountRepository.findById(userId).orElseThrow(EntityNotFoundException::new);
             return UserAccount.UserAccountDto.from(userAccount);
         } catch (Exception e) {
             throw new IllegalArgumentException("회원정보 조회에 실패했습니다");
@@ -108,11 +89,12 @@ public class UserService {
     }
 
     public void deleteUserAccount(String userId) {
-        try {
-            userAccountRepository.deleteById(userId);
-        } catch (Exception e) {
-            throw new IllegalArgumentException("회원정보 삭제에 실패했습니다");
-        }
+       if(userAccountRepository.existsById(userId)){
+           userAccountRepository.deleteById(userId);
+       }
+       else{
+              throw new EntityNotFoundException("존재하지 않는 회원입니다");
+       }
     }
     @Transactional(readOnly = true)
     public boolean isExistUser(String userId) {
@@ -145,24 +127,15 @@ public class UserService {
         return articleDtos;
     }
 
-    public String saveUserAccountWithoutProfile(UserAccount.SignupDto user) throws IOException {
-        SaveFile defaultImg = fileRepository.findByFileName("default.png");
-        String password = user.getPassword1();
+    public String saveUserAccountWithoutProfile(UserAccount.SignupDto user) {
+        SaveFile defaultImg = fileRepository.findByFileName("default.jpg");
         UserAccount account = userAccountRepository.save(user.toEntity());
-        account.setUserPassword(new BCryptPasswordEncoder().encode(password));
+        account.setUserPassword(new BCryptPasswordEncoder().encode(user.getPassword1()));
         if(defaultImg!=null){
             account.setProfileImg(defaultImg);
         }
         else{
-            SaveFile saveFile = SaveFile.builder()
-                    .fileName("default.jpg")
-                    .filePath(uploadPath+"/default.jpg")
-                    .fileSize(0L)
-                    .fileType("jpg")
-                    .uploadUser(account.getUserId())
-                    .build();
-            fileRepository.save(saveFile);
-            account.setProfileImg(saveFile);
+            throw new EntityNotFoundException("default.jpg 파일이 존재하지 않습니다");
         }
         return account.getUserId();
     }
